Change implementation of audio triggers
Initially, the audio trigger consisted of 11 bits code, encoding information about the events (successive square pulses). This turned out to be almost impossible to parse due to audio timing inaccuracies. We then opted for a much simpler implementation, whereby we send only a square pulse for all events, as a backup for the photodiode trigger. In the end, the photodiode triggers were never used, as the timing of audio event is too jittered. We only relied on the photodiode to parse events.